Changes after consultations
