+++ src/main/java/com/example/TravelAgency/domain/user/Address.java
@@ -1,4 +1,4 @@
-package com.example.TravelAgency.user;
+package com.example.TravelAgency.domain.user;
