+++ src/main/java/com/example/TravelAgency/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.example.TravelAgency.user;
+package com.example.TravelAgency.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
